TrainingData and Labels done

+++ AI_Project_Part_2/Set.h
@@ -16,8 +16,8 @@ using namespace std;
 class Set {
 
 private:
-	int distance;
-	int orientation;
+	float distance;
+	float orientation;
 
 public:
